Ensemble for multi-class classification
Softmax outputs a list containing the probabilities that an image belongs to each class. This class creates an object that holds a list of trained multi-class classification models and the number of classes between which they are intended to distinguish. By summing together the lists containing the probability distributions and mapping the index of the highest value element in the resulting list of sums to the class with which it corresponds, the model outputs a prediction based on the predictions of several trained models.